From IDE

+++ app/Http/Controllers/GalleryController.php
@@ -124,10 +124,10 @@ class GalleryController extends Controller {
         $mover = "herro im mover";
         $error=0;
 
-        if(!$fileTmpLoc){
+        if($error!=1 && !$fileTmpLoc){
             $mover="Please select a file first";
             $error=1;
-        } else if($fileSize>2097152){
+        } else if($error!=1 && $fileSize>2097152){
             $mover="File too large 10Mb max";
             unlink($fileTmpLoc);
             $error=1;
